Split pybind11 generated code across more files.
